[ALLI-7345] Refactor payment handlers to proper services with a plugin manager.
Also moves connectors to their own namespace.
