+++ module/Finna/src/Finna/OnlinePayment/Handler/AbstractBaseFactory.php
@@ -1,11 +1,10 @@
 <?php
 /**
- * Online payment factory.
+ * Default factory for payment handlers.
  *
  * PHP version 7
  *
- * Copyright (C) Villanova University 2019.
- * Copyright (C) The National Library of Finland 2019.
+ * Copyright (C) The National Library of Finland 2022.
  *
  * This program is free software; you can redistribute it and/or modify
  * it under the terms of the GNU General Public License version 2,
@@ -21,13 +20,12 @@
  * Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301  USA
  *
  * @category VuFind
- * @package  Service
- * @author   Demian Katz <demian.katz@villanova.edu>
+ * @package  OnlinePayment
  * @author   Ere Maijala <ere.maijala@helsinki.fi>
  * @license  http://opensource.org/licenses/gpl-2.0.php GNU General Public License
  * @link     https://vufind.org/wiki/development Wiki
  */
-namespace Finna\OnlinePayment;
+namespace Finna\OnlinePayment\Handler;
 
 use Interop\Container\ContainerInterface;
 use Interop\Container\Exception\ContainerException;
@@ -36,16 +34,15 @@ use Laminas\ServiceManager\Exception\ServiceNotFoundException;
 use Laminas\ServiceManager\Factory\FactoryInterface;
 
 /**
- * Online payment factory.
+ * Default factory for payment handlers.
  *
  * @category VuFind
- * @package  Service
- * @author   Demian Katz <demian.katz@villanova.edu>
+ * @package  OnlinePayment
  * @author   Ere Maijala <ere.maijala@helsinki.fi>
  * @license  http://opensource.org/licenses/gpl-2.0.php GNU General Public License
  * @link     https://vufind.org/wiki/development Wiki
  */
-class OnlinePaymentFactory implements FactoryInterface
+class AbstractBaseFactory implements FactoryInterface
 {
     /**
      * Create an object
@@ -59,19 +56,19 @@ class OnlinePaymentFactory implements FactoryInterface
      * @throws ServiceNotFoundException if unable to resolve the service.
      * @throws ServiceNotCreatedException if an exception is raised when
      * creating a service.
-     * @throws ContainerException if any other error occurs
+     * @throws ContainerException&\Throwable if any other error occurs
      */
     public function __invoke(
         ContainerInterface $container,
         $requestedName,
         array $options = null
     ) {
-        if (!empty($options)) {
-            throw new \Exception('Unexpected options passed to factory.');
-        }
+        $tableManager = $container->get(\VuFind\Db\Table\PluginManager::class);
         return new $requestedName(
-            $container->get(\Finna\OnlinePayment\Handler\PluginManager::class),
-            $container->get(\VuFind\Config\PluginManager::class)->get('datasources')
+            $container->get(\VuFindHttp\HttpService::class),
+            $container->get(\VuFind\I18n\Locale\LocaleSettings::class),
+            $tableManager->get('Transaction'),
+            $tableManager->get('Fee')
         );
     }
 }
