+++ module/Finna/src/Finna/OnlinePayment/Handler/Connector/Paytrail/PaytrailPaymentAPI/Customer.php
@@ -26,7 +26,7 @@
  * @link     http://vufind.org/wiki/vufind2:developer_manual Wiki
  * @link     http://docs.paytrail.com/ Paytrail API documentation
  */
-namespace Finna\OnlinePayment\Paytrail\PaytrailPaymentAPI;
+namespace Finna\OnlinePayment\Handler\Connector\Paytrail\PaytrailPaymentAPI;
 
 use Paytrail\SDK\Exception\ValidationException;
