+++ module/Finna/src/Finna/OnlinePayment/Handler/Paytrail.php
@@ -1,10 +1,10 @@
 <?php
 /**
- * Paytrail Payment API handler
+ * Paytrail E2 (legacy) payment handler
  *
  * PHP version 7
  *
- * Copyright (C) The National Library of Finland 2022.
+ * Copyright (C) The National Library of Finland 2014-2022.
  *
  * This program is free software; you can redistribute it and/or modify
  * it under the terms of the GNU General Public License version 2,
@@ -21,43 +21,39 @@
  *
  * @category VuFind
  * @package  OnlinePayment
+ * @author   Leszek Manicki <leszek.z.manicki@helsinki.fi>
+ * @author   Samuli Sillanpää <samuli.sillanpaa@helsinki.fi>
  * @author   Ere Maijala <ere.maijala@helsinki.fi>
  * @license  http://opensource.org/licenses/gpl-2.0.php GNU General Public License
  * @link     http://vufind.org/wiki/vufind2:developer_manual Wiki
  * @link     http://docs.paytrail.com/ Paytrail API documentation
  */
-namespace Finna\OnlinePayment;
+namespace Finna\OnlinePayment\Handler;
 
-use Finna\OnlinePayment\Paytrail\PaytrailPaymentAPI\Customer;
-use Paytrail\SDK\Model\CallbackUrl;
-use Paytrail\SDK\Model\Item;
-use Paytrail\SDK\Request\PaymentRequest;
-use Paytrail\SDK\Util\Signature;
+use Finna\OnlinePayment\Handler\Connector\Paytrail\PaytrailE2;
 
 /**
- * Paytrail Payment API handler module.
+ * Paytrail E2 (legacy) payment handler module.
  *
  * @category VuFind
  * @package  OnlinePayment
- * @author   Ere Maijala <ere.maijala@helsinki.fi>
+ * @author   Leszek Manicki <leszek.z.manicki@helsinki.fi>
+ * @author   Samuli Sillanpää <samuli.sillanpaa@helsinki.fi>
  * @license  http://opensource.org/licenses/gpl-2.0.php GNU General Public License
  * @link     http://vufind.org/wiki/vufind2:developer_manual Wiki
  * @link     http://docs.paytrail.com/ Paytrail API documentation
  */
-class PaytrailPaymentAPI extends AbstractHandler
+class Paytrail extends AbstractBase
 {
-    public const PAYMENT_SUCCESS = 'success';
-    public const PAYMENT_CANCEL = 'cancel';
-
     /**
      * Mappings from VuFind language codes to Paytrail
      *
      * @var array
      */
     protected $languageMap = [
-        'fi' => 'FI',
-        'sv' => 'SV',
-        'en' => 'EN'
+        'fi' => 'fi_FI',
+        'sv' => 'sv_SE',
+        'en' => 'en_US'
     ];
 
     /**
@@ -91,6 +87,8 @@ class PaytrailPaymentAPI extends AbstractHandler
         $patronId = $patron['cat_username'];
         $transactionId = $this->generateTransactionId($patronId);
 
+        $module = $this->initPaytrail($transactionId, $currency);
+
         $returnUrl = $this->addQueryParams(
             $returnBaseUrl,
             [$paymentParam => $transactionId]
@@ -100,45 +98,60 @@ class PaytrailPaymentAPI extends AbstractHandler
             [$paymentParam => $transactionId]
         );
 
-        $returnUrls = (new CallbackUrl())
-            ->setSuccess($returnUrl)
-            ->setCancel($returnUrl);
+        $module->setUrls($returnUrl, $returnUrl, $notifyUrl);
+        $module->setOrderNumber($transactionId);
+        $module->setCurrency($currency);
 
-        $callbackUrls = (new CallbackUrl())
-            ->setSuccess($notifyUrl)
-            ->setCancel($notifyUrl);
+        if (!empty($this->config->paymentDescription)) {
+            $module->setMerchantDescription(
+                $this->config->paymentDescription . " - $patronId"
+            );
+        } else {
+            $module->setMerchantDescription($patronId);
+        }
 
-        $names = $this->extractUserNames($user);
-        $customer = (new Customer())
-            ->setFirstName($names['firstname'] ?: 'ei tietoa')
-            ->setLastName($names['lastname'] ?: 'ei tietoa')
-            ->setEmail(trim($user->email));
+        $lastname = trim($user->lastname);
+        if (!empty($user->firstname)) {
+            $module->setFirstName(trim($user->firstname));
+        } else {
+            // We don't have both names separately, try to extract first name from
+            // last name.
+            if (strpos($lastname, ',') > 0) {
+                // Lastname, Firstname
+                [$lastname, $firstname] = explode(',', $lastname, 2);
+            } else {
+                // First Middle Last
+                if (preg_match('/^(.*) (.*?)$/', $lastname, $matches)) {
+                    $firstname = $matches[1];
+                    $lastname = $matches[2];
+                } else {
+                    $firstname = '';
+                }
+            }
+            $lastname = trim($lastname);
+            $firstname = trim($firstname);
+            $module->setFirstName(empty($firstname) ? 'ei tietoa' : $firstname);
+        }
+        $module->setLastName(empty($lastname) ? 'ei tietoa' : $lastname);
 
-        $language = $this->languageMap[$this->getCurrentLanguageCode()] ?? 'EN';
-        $paymentRequest = (new PaymentRequest())
-            ->setStamp($transactionId)
-            ->setRedirectUrls($returnUrls)
-            ->setCallbackUrls($callbackUrls)
-            ->setReference("$transactionId - $patronId")
-            ->setCurrency($currency)
-            ->setLanguage($language)
-            ->setAmount($amount + $transactionFee)
-            ->setCustomer($customer);
-        // Payment description in $this->config->paymentDescription is not supported
+        $email = trim($user->email);
+        if ($email) {
+            $module->setEmail($email);
+        }
 
-        if (isset($this->config->productCode)
-            || isset($this->config->transactionFeeProductCode)
-            || isset($this->config->productCodeMappings)
-            || isset($this->config->organizationProductCodeMappings)
+        if (!isset($this->config->productCode)
+            && !isset($this->config->transactionFeeProductCode)
+            && !isset($this->config->productCodeMappings)
+            && !isset($this->config->organizationProductCodeMappings)
         ) {
-            // Map fines to items:
+            $module->setTotalAmount($amount + $transactionFee);
+        } else {
             $productCode = !empty($this->config->productCode)
                 ? $this->config->productCode : '';
             $productCodeMappings = $this->getProductCodeMappings();
             $organizationProductCodeMappings
                 = $this->getOrganizationProductCodeMappings();
 
-            $items = [];
             foreach ($fines as $fine) {
                 $fineType = $fine['fine'] ?? '';
                 $fineOrg = $fine['organization'] ?? '';
@@ -154,7 +167,7 @@ class PaytrailPaymentAPI extends AbstractHandler
                     $code = $organizationProductCodeMappings[$fineOrg]
                         . ($productCodeMappings[$fineType] ?? '');
                 }
-                $code = mb_substr($code, 0, 100, 'UTF-8');
+                $code = mb_substr($code, 0, 16, 'UTF-8');
 
                 $fineDesc = '';
                 if (!empty($fineType)) {
@@ -168,46 +181,44 @@ class PaytrailPaymentAPI extends AbstractHandler
                     }
                 }
                 if (!empty($fine['title'])) {
-                    $title = mb_substr(
-                        $fine['title'],
-                        0,
-                        1000 - 4 - mb_strlen($fineDesc, 'UTF-8'),
-                        'UTF-8'
-                    );
-                    $fineDesc .= " ($title)";
+                    $fineDesc .= ' ('
+                        . mb_substr(
+                            $fine['title'],
+                            0,
+                            255 - 4 - strlen($fineDesc),
+                            'UTF-8'
+                        ) . ')';
                 }
-                $item = (new Item())
-                    ->setDescription($fineDesc)
-                    ->setProductCode($code)
-                    ->setUnitPrice($fine['balance'])
-                    ->setUnits(1)
-                    ->setVatPercentage(0);
-                $items[] = $item;
+                $module->addProduct(
+                    $fineDesc,
+                    $code,
+                    1,
+                    $fine['balance'],
+                    0,
+                    PaytrailE2::TYPE_NORMAL
+                );
             }
             if ($transactionFee) {
                 $code = $this->config->transactionFeeProductCode ?? $productCode;
-                $item = (new Item())
-                    ->setDescription(
-                        'Palvelumaksu / Serviceavgift / Transaction fee'
-                    )
-                    ->setProductCode($code)
-                    ->setUnitPrice($transactionFee)
-                    ->setUnits(1)
-                    ->setVatPercentage(0);
-                $items[] = $item;
+                $module->addProduct(
+                    'Palvelumaksu / Serviceavgift / Transaction fee',
+                    $code,
+                    1,
+                    $transactionFee,
+                    0,
+                    PaytrailE2::TYPE_HANDLING
+                );
             }
-            $paymentRequest->setItems($items);
         }
 
         try {
-            $paymentResponse = $this->initClient()->createPayment($paymentRequest);
+            $formData = $module->createPaymentFormData();
         } catch (\Exception $e) {
-            $request = json_encode($paymentRequest, JSON_PRETTY_PRINT);
             $this->logPaymentError(
-                'exception sending payment: ' . $e->getMessage(),
-                compact('user', 'patron', 'fines', 'request')
+                'error creating payment form data: ' . $e->getMessage(),
+                compact('user', 'patron', 'fines', 'module')
             );
-            return '';
+            return false;
         }
 
         $success = $this->createTransaction(
@@ -221,11 +232,13 @@ class PaytrailPaymentAPI extends AbstractHandler
             $fines
         );
         if (!$success) {
-            return false;
+            return '';
         }
 
-        $this->redirectToPayment($paymentResponse->getHref());
+        $paytrailUrl = !empty($this->config->e2url) ? $this->config->e2url
+            : 'https://payment.paytrail.com/e2';
 
+        $this->redirectToPaymentForm($paytrailUrl, $formData);
         return '';
     }
 
@@ -252,35 +265,28 @@ class PaytrailPaymentAPI extends AbstractHandler
             ];
         }
 
-        $status = $params['checkout-status'];
-        switch ($status) {
-        case 'ok':
-            $transaction->setPaid();
+        $status = $params['STATUS'];
+        $timestamp = $params['TIMESTAMP'];
+
+        if ('PAID' === $status) {
+            $transaction->setPaid($timestamp);
             return [
                 'success' => true,
                 'markFeesAsPaid' => true,
                 'message' => ''
             ];
-        case 'fail':
+        } elseif ('CANCELLED' === $status) {
             $transaction->setCanceled();
             return [
                 'success' => true,
                 'markFeesAsPaid' => false,
                 'message' => 'online_payment_canceled'
             ];
-        case 'new':
-        case 'pending':
-        case 'delayed':
-            return [
-                'success' => true,
-                'markFeesAsPaid' => false,
-                'message' => 'online_payment_in_progress'
-            ];
         }
 
-        $transaction->setRegistrationFailed("unknown status $status");
+        $this->logPaymentError("unknown status $status");
         return [
-            'success' => true,
+            'success' => false,
             'markFeesAsPaid' => false,
             'message' => 'online_payment_failed'
         ];
@@ -291,54 +297,61 @@ class PaytrailPaymentAPI extends AbstractHandler
      *
      * @param Laminas\Http\Request $request Request
      *
-     * @return array|false
+     * @return array
      */
-    public function getPaymentResponseParams($request)
+    protected function getPaymentResponseParams($request)
     {
-        $params = $request->getQuery()->toArray();
+        if (!($module = $this->initPaytrail())) {
+            return false;
+        }
+
+        $params = array_merge(
+            $request->getQuery()->toArray(),
+            $request->getPost()->toArray()
+        );
 
         $required = [
-            'checkout-reference',
-            'checkout-stamp',
-            'checkout-status',
-            'signature'
+            'ORDER_NUMBER', 'PAYMENT_ID', 'TIMESTAMP', 'STATUS', 'RETURN_AUTHCODE'
         ];
 
         foreach ($required as $name) {
-            if (empty($params[$name])) {
+            if (!isset($params[$name])) {
                 $this->logPaymentError(
-                    "missing or empty parameter $name in payment response",
+                    "missing parameter $name in payment response",
                     compact('params')
                 );
                 return false;
             }
         }
 
-        // Validate the parameters:
-        try {
-            Signature::validateHmac(
-                $params,
-                '',
-                $params['signature'],
-                $this->config['secret'] ?? ''
-            );
-        } catch (\Exception $e) {
+        $success = $module->validateRequest(
+            $params['ORDER_NUMBER'],
+            $params['PAYMENT_ID'],
+            $params['TIMESTAMP'],
+            $params['STATUS'],
+            $params['RETURN_AUTHCODE']
+        );
+        if (!$success) {
             $this->logPaymentError(
-                'parameter signature validation failed',
-                compact('params')
+                'error processing response: invalid checksum',
+                compact('request', 'params')
             );
-            return false;
+            return [
+                'success' => false,
+                'markFeesAsPaid' => false,
+                'error' => 'online_payment_failed'
+            ];
         }
 
         return $params;
     }
 
     /**
-     * Initialize the Paytrail client
+     * Initialize the Paytrail module
      *
-     * @return Paytrail\PaytrailPaymentAPI\Client
+     * @return PaytrailE2
      */
-    protected function initClient(): Paytrail\PaytrailPaymentAPI\Client
+    protected function initPaytrail()
     {
         foreach (['merchantId', 'secret'] as $req) {
             if (!isset($this->config[$req])) {
@@ -347,13 +360,57 @@ class PaytrailPaymentAPI extends AbstractHandler
             }
         }
 
-        $client = new Paytrail\PaytrailPaymentAPI\Client(
+        return new PaytrailE2(
             $this->config->merchantId,
             $this->config->secret,
-            'Finna'
+            $this->languageMap[$this->getCurrentLanguageCode()] ?? 'en_US'
         );
-        $client->setHttpService($this->http);
-        $client->setLogger($this->logger);
-        return $client;
+    }
+
+    /**
+     * Redirect to payment handler.
+     *
+     * @param string $url      URL
+     * @param array  $formData Form fields
+     *
+     * @return void
+     */
+    protected function redirectToPaymentForm($url, $formData)
+    {
+        // Output a minimal form and submit it automatically
+        $formFields = '';
+        foreach ($formData as $key => $value) {
+            $formFields .= '<input type="hidden" name="' . htmlentities($key)
+                . '" value="' . htmlentities($value) . '">';
+        }
+        $lang = $this->getCurrentLanguageCode();
+        $title = $this->translator->translate('online_payment_go_to_pay');
+        $title = str_replace('%%amount%%', '', $title);
+        $jsRequired = $this->translator->translate('Please enable JavaScript.');
+        echo <<<EOT
+<!DOCTYPE html>
+<html lang="$lang">
+<head>
+    <meta http-equiv="X-UA-Compatible" content="IE=edge"/>
+    <meta name="viewport" content="width=device-width,initial-scale=1.0"/>
+    <meta http-equiv="Content-Type" content="text/html; charset=UTF-8">
+    <title>$title</title>
+    <script>
+        document.addEventListener('DOMContentLoaded', function() {
+            document.getElementById('paytrail-form').submit();
+        });
+    </script>
+</head>
+<body>
+    <noscript>
+        $jsRequired
+    </noscript>
+    <form id="paytrail-form" action="$url" method="POST">
+        $formFields
+    </form>
+</body>
+</html>
+EOT;
+        exit();
     }
 }
