+++ module/Finna/src/Finna/OnlinePayment/Handler/PluginManager.php
@@ -1,0 +1,75 @@
+<?php
+/**
+ * Online payment handler plugin manager
+ *
+ * PHP version 7
+ *
+ * Copyright (C) The National Library of Finland 2022.
+ *
+ * This program is free software; you can redistribute it and/or modify
+ * it under the terms of the GNU General Public License version 2,
+ * as published by the Free Software Foundation.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU General Public License for more details.
+ *
+ * You should have received a copy of the GNU General Public License
+ * along with this program; if not, write to the Free Software
+ * Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301  USA
+ *
+ * @category VuFind
+ * @package  OnlinePayment
+ * @author   Ere Maijala <ere.maijala@helsinki.fi>
+ * @license  http://opensource.org/licenses/gpl-2.0.php GNU General Public License
+ * @link     https://vufind.org/wiki/development Wiki
+ */
+namespace Finna\OnlinePayment\Handler;
+
+/**
+ * Online payment handler plugin manager
+ *
+ * @category VuFind
+ * @package  OnlinePayment
+ * @author   Ere Maijala <ere.maijala@helsinki.fi>
+ * @license  http://opensource.org/licenses/gpl-2.0.php GNU General Public License
+ * @link     https://vufind.org/wiki/development Wiki
+ */
+class PluginManager extends \VuFind\ServiceManager\AbstractPluginManager
+{
+    /**
+     * Default plugin aliases.
+     *
+     * @var array
+     */
+    protected $aliases = [
+        'CPU' => CPU::class,
+        'Paytrail' => Paytrail::class,
+        'PaytrailPaymentAPI' => PaytrailPaymentAPI::class,
+        'TurkuPayment' => TurkuPayment::class,
+    ];
+
+    /**
+     * Default plugin factories.
+     *
+     * @var array
+     */
+    protected $factories = [
+        CPU::class => AbstractBaseFactory::class,
+        Paytrail::class => AbstractBaseFactory::class,
+        PaytrailPaymentAPI::class => AbstractBaseFactory::class,
+        TurkuPayment::class => AbstractBaseFactory::class,
+    ];
+
+    /**
+     * Return the name of the base class or interface that plug-ins must conform
+     * to.
+     *
+     * @return string
+     */
+    protected function getExpectedInterface()
+    {
+        return AbstractBase::class;
+    }
+}
